projects/lab01/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huepy==1.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
instabot==0.117.0
jiter==0.10.0
mock==5.2.0
moviepy==2.2.1
numpy==2.2.6
openai==1.82.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
proglog==0.1.12
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.7
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
